requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.2.7
future==0.18.2
idna==2.8
line-bot-sdk==1.15.0
pyaml==19.4.1
pygraphviz==1.5
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
termcolor==1.1.0
transitions==0.7.1
urllib3==1.25.7
